requirements.txt
# Apache Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# Database drivers
psycopg2-binary==2.9.9
SQLAlchemy==1.4.51

# Dask for distributed computing
dask[complete]==2023.12.1
distributed==2023.12.1
dask-sql==2023.12.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# API requests
requests==2.31.0
aiohttp==3.9.1

# Moscow Exchange API
apimoex==1.3.0

# Data validation
pydantic==2.5.3

# Configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Date/Time utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
